requirements.txt
boto3
opensearch-py 
requests-aws4auth
python-dotenv